Renaming package to "com.mangoblogger.app". Added Fragments code.
