+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 25
     buildToolsVersion "25.0.2"
     defaultConfig {
-        applicationId "app.com.example.karthik.mangoblogger"
+        applicationId "com.mangoblogger.app"
         minSdkVersion 9
         targetSdkVersion 25
         versionCode 1
