+++ app/src/main/java/com/mangoblogger/app/AnalyticsTermsFragment.java
@@ -1,0 +1,20 @@
+package com.mangoblogger.app;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+/**
+ * Created by karthikprasad on 7/29/17.
+ */
+
+public class AnalyticsTermsFragment extends Fragment {
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_analytics_terms, container, false);
+    }
+}
