+++ app/src/main/java/com/mangoblogger/app/MainActivity.java
@@ -1,15 +1,14 @@
-package app.com.example.karthik.mangoblogger;
+package com.mangoblogger.app;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.design.widget.BottomNavigationView;
+import android.support.v4.app.Fragment;
 import android.support.v7.app.AppCompatActivity;
 import android.view.MenuItem;
-import android.widget.TextView;
+
 
 public class MainActivity extends AppCompatActivity {
-
-    private TextView mTextMessage;
 
     private BottomNavigationView.OnNavigationItemSelectedListener mOnNavigationItemSelectedListener
             = new BottomNavigationView.OnNavigationItemSelectedListener() {
@@ -18,13 +17,16 @@ public class MainActivity extends AppCompatActivity {
         public boolean onNavigationItemSelected(@NonNull MenuItem item) {
             switch (item.getItemId()) {
                 case R.id.navigation_home:
-                    mTextMessage.setText(R.string.analytics_terms_title);
+                    //inflate analytics fragment
+                    changeFragment(0);
                     return true;
                 case R.id.navigation_dashboard:
-                    mTextMessage.setText(R.string.ux_terms_title);
+                    //inflate ux fragment
+                    changeFragment(1);
                     return true;
                 case R.id.navigation_notifications:
-                    mTextMessage.setText(R.string.about_title);
+                    //inflate about fragment
+                    changeFragment(2);
                     return true;
             }
             return false;
@@ -36,10 +38,31 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        mTextMessage = (TextView) findViewById(R.id.message);
+        changeFragment(0);
         BottomNavigationView navigation = (BottomNavigationView) findViewById(R.id.navigation);
         navigation.setOnNavigationItemSelectedListener(mOnNavigationItemSelectedListener);
+    }
+
+    /**
+     * To load fragments for sample
+     *
+     * @param position menu index
+     */
+    private void changeFragment(int position) {
+
+        Fragment newFragment = null;
+
+        if (position == 0) {
+            newFragment = new AnalyticsTermsFragment();
+        } else if (position == 1) {
+            newFragment = new UxTermsFragment();
+        } else {
+            newFragment = new AboutFragment();
+        }
+
+        getSupportFragmentManager().beginTransaction().replace(
+                R.id.container, newFragment)
+                .commit();
     }
 
 }
